prepare model for new storage

+++ internal/web/db/cache.go
@@ -6,11 +6,11 @@ import (
 
 func (db *Config) append(e *v1.Model) {
 	hit(e)
-	if cached, ok := db.cache.Get(e.Id); ok {
+	if cached, ok := db.cache.Get(uint64(e.Id)); ok {
 		update(cached, e)
 		db.ts.Save(e)
 		return
 	}
 	db.ts.Save(e)
-	db.cache.Add(e.Id, e)
+	db.cache.Add(uint64(e.Id), e)
 }
